refactor: ChannelDialog open 속성을 custom hook에서 관리
Dialog 외부에서 handleClickOpen 콜백이 필요하다 생각되어 custom hook으로 관리한다.
